refactor(cluster-adaptor-std): reorganize module boundaries
