+++ modules/cluster-adaptor-std/src/cluster_provider/ecs_poller_error.rs
@@ -1,0 +1,10 @@
+//! Error type for ECS polling operations.
+
+use std::string::String;
+
+/// Error type for ECS polling operations.
+#[derive(Debug)]
+pub enum EcsPollerError {
+  /// API call failed.
+  ApiCall(String),
+}
