+++ modules/cluster-adaptor-std/src/membership.rs
@@ -1,0 +1,15 @@
+//! Tokio-backed membership and gossip adaptors.
+
+mod gossip_wire_delta_v1;
+mod gossip_wire_node_record;
+mod membership_coordinator_driver;
+mod tokio_gossip_transport;
+mod tokio_gossip_transport_config;
+mod tokio_gossiper;
+mod tokio_gossiper_config;
+
+pub use membership_coordinator_driver::MembershipCoordinatorDriver;
+pub use tokio_gossip_transport::TokioGossipTransport;
+pub use tokio_gossip_transport_config::TokioGossipTransportConfig;
+pub use tokio_gossiper::TokioGossiper;
+pub use tokio_gossiper_config::TokioGossiperConfig;
