+++ modules/cluster-adaptor-std/src/membership/gossip_wire_node_record.rs
