+++ modules/cluster-adaptor-std/src/pub_sub.rs
@@ -1,0 +1,5 @@
+//! Cluster publish/subscribe delivery adaptors.
+
+mod pub_sub_delivery_actor;
+
+pub use pub_sub_delivery_actor::PubSubDeliveryActor;
